LayoutTests/js/kde/md5-2-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS hexMD5('kde') is '186cf28b76f2264e9fea8fcf91cb4f5d'
PASS successfullyParsed is true

TEST COMPLETE
